requirements.txt
joblib==0.13.2
numpy==1.16.5
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
pandas==0.25.1
Pillow==7.0.0
scikit-image==0.16.2
scikit-learn==0.21.2

